+++ src/common/types/popup.ts
@@ -18,7 +18,7 @@ export type Popups = {
     groupName: PopupItem
     bindGroup: PopupItem
     chooseEmoji: PopupItem<{}, { emo: string }>
-    chooseImageIcon: PopupItem<{}, { url: string }>
+    chooseImageIcon: PopupItem<{}, { url: string | null }>
     newPassword: PopupItem<{}, { newPass: string }>
     password: PopupItem<PasswordData>
     linkMenuView: PopupItem<{ link: Link }>