+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - 🧑‍💻 Improve styles for tab icons in a group.
 - 🧑‍💻 Added 1 more "Pin" icon to choose for a group.
 - 🐛 Fixed validation error message text when password is empty. It was `password_empty`.
+- 🐛 Fixed not working "Cancel" button when choosing group image icon.
 
 ## v1.15.0 (2025-10-19)
 - ✨ Group icon on a group page is now clickable. It will open "Choose Icon" interface [#53](https://github.com/tab-guardian/tab-guardian/issues/53).